Upload mainnet v1.0.2 source
Former-commit-id: d1201d5c49324ac979fdeb1f28bbed54916f6966

+++ libraries/fc/include/fc/io/varint.hpp
@@ -4,7 +4,7 @@
 namespace fc {
 
 struct unsigned_int {
-    unsigned_int( uint32_t v = 0 ):value(v){}
+    unsigned_int( uint64_t v = 0 ):value(v){}
 
     template<typename T>
     unsigned_int( T v ):value(v){}
@@ -17,22 +17,22 @@ struct unsigned_int {
 
     unsigned_int& operator=( int32_t v ) { value = v; return *this; }
     
-    uint32_t value;
+    uint64_t value;
 
-    friend bool operator==( const unsigned_int& i, const uint32_t& v )     { return i.value == v; }
-    friend bool operator==( const uint32_t& i, const unsigned_int& v )     { return i       == v.value; }
+    friend bool operator==( const unsigned_int& i, const uint64_t& v )     { return i.value == v; }
+    friend bool operator==( const uint64_t& i, const unsigned_int& v )     { return i       == v.value; }
     friend bool operator==( const unsigned_int& i, const unsigned_int& v ) { return i.value == v.value; }
 
-    friend bool operator!=( const unsigned_int& i, const uint32_t& v )     { return i.value != v; }
-    friend bool operator!=( const uint32_t& i, const unsigned_int& v )     { return i       != v.value; }
+    friend bool operator!=( const unsigned_int& i, const uint64_t& v )     { return i.value != v; }
+    friend bool operator!=( const uint64_t& i, const unsigned_int& v )     { return i       != v.value; }
     friend bool operator!=( const unsigned_int& i, const unsigned_int& v ) { return i.value != v.value; }
 
-    friend bool operator<( const unsigned_int& i, const uint32_t& v )      { return i.value < v; }
-    friend bool operator<( const uint32_t& i, const unsigned_int& v )      { return i       < v.value; }
+    friend bool operator<( const unsigned_int& i, const uint64_t& v )      { return i.value < v; }
+    friend bool operator<( const uint64_t& i, const unsigned_int& v )      { return i       < v.value; }
     friend bool operator<( const unsigned_int& i, const unsigned_int& v )  { return i.value < v.value; }
 
-    friend bool operator>=( const unsigned_int& i, const uint32_t& v )     { return i.value >= v; }
-    friend bool operator>=( const uint32_t& i, const unsigned_int& v )     { return i       >= v.value; }
+    friend bool operator>=( const unsigned_int& i, const uint64_t& v )     { return i.value >= v; }
+    friend bool operator>=( const uint64_t& i, const unsigned_int& v )     { return i       >= v.value; }
     friend bool operator>=( const unsigned_int& i, const unsigned_int& v ) { return i.value >= v.value; }
 };
 
@@ -95,7 +95,7 @@ namespace std
        public:
          size_t operator()(const fc::signed_int &a) const 
          {
-            return std::hash<uint32_t>()(a.value);
+            return std::hash<uint64_t>()(a.value);
          }
    };
 }
